src/main/java/ru/sfedu/sprintspherepk/util/XmlUtil.java
package ru.sfedu.sprintspherepk.util;

import org.w3c.dom.Document;
import javax.xml.parsers.DocumentBuilder;
import javax.xml.parsers.DocumentBuilderFactory;
import java.io.File;

public class XmlUtil {
    public static Document loadXml(String path) {
        try {
            File xmlFile = new File(path);
            DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
            DocumentBuilder builder = factory.newDocumentBuilder();
            return builder.parse(xmlFile);
        } catch (Exception e) {
            throw new RuntimeException("Failed to load XML file: " + e.getMessage());
        }
    }
}